Azure Functions/fingrid_api_real_time_data_azure_function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.27.1
azure-storage-blob==12.8.1